seperating out classes for todo app
